bug: Start-Transcript noisy
when using GitlabCli inside a session transcription, the output would be noisy (AF), e.g.
PS>TerminatingError(): “The pipeline has been stopped.”
>> TerminatingError(): “The pipeline has been stopped.”
PS>TerminatingError(): “The pipeline has been stopped.”
>> TerminatingError(): “The pipeline has been stopped.”
...
This output would not be visible in the console, only in the transcript output file.

Isolated the source to inside Invoke-GitlabApi where it is resolving which site to use
$(Get-GitlabConfiguration | Select-Object -ExpandProperty Sites | Select -First 1 -ExpandProperty Url)
results in the log noise

$(Get-GitlabConfiguration | Select-Object -ExpandProperty Sites)[0].Url
yields the same result, no noise

still not sure WHY, but sidestepping the problem as the intent of the code is for there only to ever be 0 or 1 match(es)