Fixes #{insertissuenumber}
The unit tests were created to vet the regex updates as they can be very fiddly.

Fixed issue with getting 404 when using any GitlabCli commands that interrogate the current folder for git information. Projects can have a `.` in their path and this wasn't accomodated for in the current regex
Got mired in not capturing `.git` when it's there. There are two formats for git repo urls that aren't http based. These can have either .git or not on the end when running `git config --get remote.origin.url`

For some reason I wasn't able to figure that part out. Ended up with a best effort just removing .git from the end.

To invoke the tests run `Invoke-Pester` in the root of the repo, it should find them automatically.